# docs/swagger.yaml
swagger: '2.0'
info:
  title: Open edX API
  description: APIs for access to Open edX information
  contact:
    email: oscm@edx.org
  version: v1
basePath: /
consumes:
  - application/json
produces:
  - application/json
securityDefinitions:
  Basic:
    type: basic
security:
  - Basic: []
paths:
...
  /api/bookmarks/v1/bookmarks/:
    get:
      operationId: api_bookmarks_v1_bookmarks_list
      summary: Get a paginated list of bookmarks for a user.
      description: "The list can be filtered by passing parameter \"course_id=<course_id>\"\
        \nto only include bookmarks from a particular course.\n\nThe bookmarks are\
        \ always sorted in descending order by creation date.\n\nEach page in the\
        \ list contains 10 bookmarks by default. The page\nsize can be altered by\
        \ passing parameter \"page_size=<page_size>\".\n\nTo include the optional\
        \ fields pass the values in \"fields\" parameter\nas a comma separated list.\
        \ Possible values are:\n\n* \"display_name\"\n* \"path\"\n\n# Example Requests\n\
        \nGET /api/bookmarks/v1/bookmarks/?course_id={course_id1}&fields=display_name,path\n\
        \n# Response Values\n\n* count: The number of bookmarks in a course.\n\n*\
        \ next: The URI to the next page of bookmarks.\n\n* previous: The URI to the\
        \ previous page of bookmarks.\n\n* num_pages: The number of pages listing\
        \ bookmarks.\n\n* results:  A list of bookmarks returned. Each collection\
        \ in the list\n  contains these fields.\n\n    * id: String. The identifier\
        \ string for the bookmark: {user_id},{usage_id}.\n\n    * course_id: String.\
        \ The identifier string of the bookmark's course.\n\n    * usage_id: String.\
        \ The identifier string of the bookmark's XBlock.\n\n    * display_name: String.\
        \ (optional) Display name of the XBlock.\n\n    * path: List. (optional) List\
        \ of dicts containing {\"usage_id\": <usage-id>, display_name:<display-name>}\n\
        \        for the XBlocks from the top of the course tree till the parent of\
        \ the bookmarked XBlock.\n\n    * created: ISO 8601 String. The timestamp\
        \ of bookmark's creation.\n\n"
      parameters:
        - name: page
          in: query
          description: A page number within the paginated result set.
          required: false
          type: integer
        - name: page_size
          in: query
          description: Number of results to return per page.
          required: false
          type: integer
        - name: course_id
          in: query
          description: The id of the course, of course
          type: string
        - name: fields
          in: query
          description: "The fields to return: display_name, path.\n"
          type: string
      responses:
        '200':
          description: ''
          schema:
            $ref: '#/definitions/Bookmark'
      tags:
        - api
    post:
      operationId: api_bookmarks_v1_bookmarks_create
      summary: Create a new bookmark for a user.
      description: "The POST request only needs to contain one parameter \"usage_id\"\
        .\n\nHttp400 is returned if the format of the request is not correct,\nthe\
        \ usage_id is invalid or a block corresponding to the usage_id\ncould not\
        \ be found.\n\n# Example Requests\n\nPOST /api/bookmarks/v1/bookmarks/\nRequest\
        \ data: {\"usage_id\": <usage-id>}\n\n"
      parameters: []
      responses:
        '201':
          description: ''
      tags:
        - api
    parameters: []
...
  /courses/{course_id}/xblock/{usage_id}/view/{view_name}:
    get:
      operationId: courses_xblock_view_read
      summary: Returns the rendered view of a given XBlock, with related resources
      description: "Returns a json object containing two keys:\n    html: The rendered\
        \ html of the view\n    resources: A list of tuples where the first element\
        \ is the resource hash, and\n        the second is the resource description"
      parameters: []
      responses:
        '200':
          description: ''
      tags:
        - courses
    parameters:
      - name: course_id
        in: path
        required: true
        type: string
      - name: usage_id
        in: path
        required: true
        type: string
      - name: view_name
        in: path
        required: true
        type: string
...
definitions:
...
  Bookmark:
    required:
      - course_id
      - usage_id
    type: object
    properties:
      id:
        title: Id
        type: string
        readOnly: true
      course_id:
        title: Course id
        type: string
      usage_id:
        title: Usage id
        type: string
      block_type:
        title: Block type
        type: string
        readOnly: true
      created:
        title: Created
        type: string
        format: date-time
        readOnly: true
...